refactor: 예약 컨트롤러와 성격이 다른 메서드를 분리
