+++ src/main/java/roomescape/domain/roomescape/controller/RoomEscapeController.java
@@ -1,0 +1,17 @@
+package roomescape.domain.roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class RoomEscapeController {
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "reservation";
+    }
+}
